Refactor: Main 컴포넌트 Section 분리 Main 페이지 section들 컴포넌트로 분리
