+++ src/pages/Main/sections/CircleCategory/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CircleCategory';
