+++ src/pages/Main/sections/OurPicksList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OurPicksList';
