+++ src/pages/Main/sections/PopularList/PopularList.style.ts
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const SectionTitle = styled.h2`
+  font-size: 28px;
+  margin-bottom: 18px;
+  font-weight: 500;
+  padding: 0 9px;
+`;
+
+export const PopularContainer = styled.div`
+  max-width: 1400px;
+  margin-bottom: 90px;
+`;
+
+export const PopularBox = styled.div`
+  display: flex;
+`;
