+++ src/pages/Main/sections/PopularList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PopularList';
