+++ src/pages/Main/sections/RecentList/RecentList.style.ts
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+
+export const RecentListWrap = styled.div`
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  width: 1400px;
+  gap: 18px;
+  margin-bottom: 30px;
+`;
+
+export const RecentTitle = styled.div`
+  display: flexbox;
+  justify-content: space-between;
+  width: 100%;
+  padding: 0 40px;
+`;
+
+export const RecentLeft = styled.p`
+  font-size: 16px;
+  font-weight: 500;
+`;
+
+export const RecentRight = styled.span`
+  font-size: 13px;
+  color: gray;
+`;
