+++ src/pages/Main/sections/RecentList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecentList';
