+++ src/pages/Main/sections/SelectionList/SelectionList.style.ts
@@ -1,0 +1,47 @@
+import styled from 'styled-components';
+
+export const SectionTitle = styled.h2`
+  font-size: 28px;
+  margin-bottom: 18px;
+  font-weight: 500;
+`;
+
+export const SVGWrapper = styled.div`
+  display: inline-block;
+  transform: translateX(10px);
+  transition: all 0.3s ease-in;
+  margin-top: 5px;
+`;
+
+export const TitleWrapper = styled.div`
+  margin-bottom: 18px;
+  width: fit-content;
+  display: flex;
+  align-items: center;
+  ${SectionTitle} {
+    display: inline-block;
+    cursor: pointer;
+    margin-bottom: 0;
+  }
+  &:hover ${SVGWrapper} {
+    transform: translateX(15px);
+  }
+`;
+
+export const SelectionsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  max-width: 1400px;
+  margin-bottom: 90px;
+`;
+
+export const SectionDesc = styled.p`
+  font-size: 16px;
+  color: gray;
+  margin-bottom: 18px;
+`;
+
+export const SelectionsWrap = styled.div`
+  display: flex;
+  gap: 18px;
+`;
