+++ src/pages/Main/sections/SelectionList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectionList';
