+++ src/pages/Main/sections/UniqueList/Unique.style.ts
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const UniqueContainer = styled.div`
+  max-width: 1400px;
+  margin-bottom: 90px;
+`;
+
+export const SectionTitle = styled.h2`
+  font-size: 28px;
+  margin-bottom: 18px;
+  font-weight: 500;
+`;
+
+export const TabContents = styled.div`
+  margin-top: 18px;
+  display: flex;
+  gap: 18px;
+  flex-wrap: wrap;
+`;
