+++ src/pages/Main/sections/UniqueList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UniqueList';
